+++ src/Release.js
@@ -1,6 +1,5 @@
 const Semver = require('semver');
 const Change = require('./Change');
-const _changelog = Symbol.for('changelog');
 
 class Release {
     constructor(version, date, description = '') {
@@ -15,22 +14,14 @@ class Release {
         this.version = version;
         this.date = date;
         this.description = description;
-        this.changes = {
-            added: [],
-            changed: [],
-            deprecated: [],
-            removed: [],
-            fixed: [],
-            security: []
-        };
-    }
-
-    set changelog(changelog) {
-        this[_changelog] = changelog;
-    }
-
-    get changelog() {
-        return this[_changelog];
+        this.changes = new Map([
+            ['added', []],
+            ['changed', []],
+            ['deprecated', []],
+            ['removed', []],
+            ['fixed', []],
+            ['security', []]
+        ]);
     }
 
     compare(release) {
@@ -50,8 +41,8 @@ class Release {
             return false;
         }
 
-        return !Object.keys(this.changes).some(
-            key => this.changes[key].length > 0
+        return Array.from(this.changes.values()).some(
+            change => change.length > 0
         );
     }
 
@@ -60,12 +51,11 @@ class Release {
             change = new Change(change);
         }
 
-        if (!Array.isArray(this.changes[type])) {
+        if (!this.changes.has(type)) {
             throw new Error('Invalid change type');
         }
 
-        this.changes[type].push(change);
-        change.release = this;
+        this.changes.get(type).push(change);
 
         return this;
     }
@@ -94,18 +84,17 @@ class Release {
         return this.addChange('security', change);
     }
 
-    toString() {
-        const url = this.changelog ? this.changelog.url : null;
+    toString(changelog) {
         let t = [];
 
         if (this.version) {
-            if (this.hasCompareLink()) {
+            if (this.hasCompareLink(changelog)) {
                 t.push(`## [${this.version}] - ${formatDate(this.date)}`);
             } else {
                 t.push(`## ${this.version} - ${formatDate(this.date)}`);
             }
         } else {
-            if (this.hasCompareLink()) {
+            if (this.hasCompareLink(changelog)) {
                 t.push('## [UNRELEASED]');
             } else {
                 t.push('## UNRELEASED');
@@ -117,58 +106,63 @@ class Release {
             t.push(this.description.trim());
         }
 
-        for (let k in this.changes) {
-            let changes = this.changes[k];
-
+        this.changes.forEach((changes, type) => {
             if (changes.length) {
                 t.push('');
-                t.push(`### ${k[0].toUpperCase()}${k.substring(1)}`);
+                t.push(`### ${type[0].toUpperCase()}${type.substring(1)}`);
                 t.push('');
                 t = t.concat(changes.map(change => change.toString()));
             }
-        }
+        });
 
         return t.join('\n').trim();
     }
 
-    getCompareLink() {
-        if (!this.hasCompareLink()) {
+    getCompareLink(changelog) {
+        if (!this.hasCompareLink(changelog)) {
             return;
-        }
-
-        const changelog = this[_changelog];
-
-        if (changelog.unreleased === this) {
-            return `[UNRELEASED]: ${changelog.url}/compare/v${changelog
-                .releases[0].version}...HEAD`;
         }
 
         const index = changelog.releases.indexOf(this);
         const next = changelog.releases[index + 1];
+
+        if (!this.version) {
+            return `[UNRELEASED]: ${changelog.url}/compare/v${next.version}...HEAD`;
+        }
 
         return `[${this
             .version}]: ${changelog.url}/compare/v${next.version}...v${this
             .version}`;
     }
 
-    hasCompareLink() {
-        const changelog = this[_changelog];
+    getLinks(changelog) {
+        const links = [];
 
+        if (!changelog.url) {
+            return links;
+        }
+
+        this.changes.forEach(changes =>
+            changes.forEach(change => {
+                change.issues.forEach(issue => {
+                    if (!links.includes(issue)) {
+                        links.push(
+                            `[#${issue}]: ${changelog.url}/issues/${issue}`
+                        );
+                    }
+                });
+            })
+        );
+
+        return links;
+    }
+
+    hasCompareLink(changelog) {
         if (!changelog || !changelog.url || !changelog.releases.length) {
             return false;
         }
 
-        if (changelog.unreleased === this) {
-            return true;
-        }
-
-        if (this.version) {
-            return (
-                changelog.releases.length > changelog.releases.indexOf(this) + 1
-            );
-        }
-
-        return false;
+        return changelog.releases.length > changelog.releases.indexOf(this) + 1;
     }
 }