+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.5.0] - 2018-05-09
+
+### Added
+
+- Allow to insert a footer at the end of the changelog
+
+### Changed
+
+- Placed the urls of the issues at bottom
+- Indent automatically the extra lines of the changes
+
+### Fixed
+
+- Parser refactoring
+
 ## [0.4.0] - 2018-03-22
 
 ### Added
@@ -13,7 +28,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
-- Use UTC date functions to fix decrement issue - [#1](https://github.com/oscarotero/keep-a-changelog/issues/1)
+- Use UTC date functions to fix decrement issue - [#1]
 
 ## [0.3.1] - 2017-12-11
 
@@ -53,6 +68,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 First version
 
+[#1]: https://github.com/oscarotero/keep-a-changelog/issues/1
+
+[0.5.0]: https://github.com/oscarotero/keep-a-changelog/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/oscarotero/keep-a-changelog/compare/v0.3.1...v0.4.0
 [0.3.1]: https://github.com/oscarotero/keep-a-changelog/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/oscarotero/keep-a-changelog/compare/v0.2.0...v0.3.0